# lib/api-zod/src/generated/types/raffle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Raffle {
  id: string;
  guild_id: string;
  channel_id: string;
  /** @nullable */
  message_id?: string | null;
  title: string;
  description: string;
  prizes: string[];
  winner_count: number;
  /** @nullable */
  ends_at?: string | null;
  started_by: string;
  status: string;
  winners: string[];
}